Iterator safety.
Move mergeSegments to a lambda so we can more easily break out of the
nested for-loops when we find a merge.  We really need to start over
again at that point as we may have invalidated our connectivity items
iterator.

This also means we don't have to queue up all the track pointers, which
should get us a _little_ bit of the performance back.

Fixes https://gitlab.com/kicad/code/kicad/issues/13287